Move application directory.
