+++ app/components/application/demos/emoji-react.js
